requirements.txt
openai
groq
yfinance
streamlit
python-dotenv
requests
pandas